First Commit
A Real time Messaging App which allows users to create accounts and chat with each other in real-time. The website is developed using MERN stack, which includes MongoDB, Express, React, and Node.js. Socket.IO is used for real-time communication between the users. Redux Toolkit is used for state management, and Tailwind CSS is used for UI